+++ src/modules/layerTree/tests/unit/components/LayerTreeNode.spec.js
@@ -157,6 +157,13 @@ describe("src/modules/layerTree/components/LayerTreeNode.vue", () => {
                                 name: () => "Contactname",
                                 type: () => "contact"
                             }
+                        },
+                        ResizeHandle: {
+                            namespaced: true,
+                            getters: {
+                                mainMenuWidth: () => 0,
+                                secondaryMenuWidth: () => 0
+                            }
                         }
                     }
                 },